Split mobile devices into its own component
- it makes more sense as a separate thing
- the correct permission for mobile devices is not READ_ACL but
  READ_MOBILE_DEVICES
- this commit doesn't fix remove/rename actions which are scoped to the
  mobile device itself. These actions are still incorrectly assumed to
  be ok if the user has READ_MOBILE_DEVICES permission in the unit
